requirements.txt
Pillow>=9.0.0
numpy>=1.21.0
viam-sdk=0.51.0